[43] feat: Add create edit profile form (#67)
* refactor: make all props in modal window not necessary

* feat: add title and button delete

* refacor: rename modal window context

* feat: add editProfile modal window info

* fix: change saveButton font to upperCase

* feat: add styles to delete Button profile

* feat: add modal window when delete profile

* refactor: change modal window names

* feat: add possibility to confirm delete profile

+++ src/components/ModalWindow/DeleteBoard/DeleteBoard.tsx
@@ -1,14 +1,18 @@
 import React, { ReactElement } from 'react';
 
 import answerImage from '../../../assets/images/answer.png';
-import { Props } from '../types/DeleteModalWindow.types';
+import { Props } from '../types/DeleteBoard.types';
 
-import styles from './DeleteModalWindow.module.scss';
+import styles from './DeleteBoard.module.scss';
 
-const DeleteModalWindow = ({ deleteBoard, toggleModalWindow }: Props): ReactElement => {
-  const deleteItem = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>): void => {
-    deleteBoard();
-    toggleModalWindow(e);
+const DeleteBoard = ({ deleteBoard, toggleModalWindow }: Props): ReactElement => {
+  const deleteItem = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>): void | undefined => {
+    if (deleteBoard !== undefined) {
+      deleteBoard();
+    }
+    if (toggleModalWindow !== undefined) {
+      toggleModalWindow(e);
+    }
   };
 
   return (
@@ -27,4 +31,4 @@ const DeleteModalWindow = ({ deleteBoard, toggleModalWindow }: Props): ReactElem
   );
 };
 
-export default DeleteModalWindow;
+export default DeleteBoard;
